Heres the code
So, one of the main limits of the program is that the server has to be initiated one or two rounds of sending UDP packets before adding on the clients. Im sure one of yall can figure out how to fix that if you want. Other than that... It runs a bit slow but, i made it that way so multi-threading and resource sharing wasnt a problem. You can probably speed it up if you want, i havnt tried. I've only tried with node 3 (as server), node 2 and rl-dual (node 4? the one w/ packet loss). Let me know when yall are planning on going through the questions

gcc -g -pthread project2bmulticastServer.c -o testrun
if you need the gcc command to compile